requirements.txt
aiohttp==3.5.4
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
ipcalc==1.99.0
multidict==4.5.2
pycparser==2.19
pyOpenSSL==19.0.0
PyYAML==5.1.1
six==1.12.0
slackclient==2.1.0
yarl==1.3.0
